test(refactor): move the RejectionTimer class to the separate folder
